# bpo-47067: Add vectorcall for gaobject