--- a/Include/internal/pycore_global_strings.h
+++ b/Include/internal/pycore_global_strings.h
@@ -156,6 +156,7 @@ struct _Py_global_strings {
         STRUCT_FOR_ID(__next__)
         STRUCT_FOR_ID(__note__)
         STRUCT_FOR_ID(__or__)
+        STRUCT_FOR_ID(__orig_class__)
         STRUCT_FOR_ID(__origin__)
         STRUCT_FOR_ID(__package__)
         STRUCT_FOR_ID(__parameters__)