--- a/Include/internal/pycore_runtime_init.h
+++ b/Include/internal/pycore_runtime_init.h
@@ -779,6 +779,7 @@ extern "C" {
                 INIT_ID(__next__), \
                 INIT_ID(__note__), \
                 INIT_ID(__or__), \
+                INIT_ID(__orig_class__), \
                 INIT_ID(__origin__), \
                 INIT_ID(__package__), \
                 INIT_ID(__parameters__), \