--- a/Misc/NEWS.d/next/Library/2022-03-20-17-15-56.bpo-47067.XXLnje.rst
+++ b/Misc/NEWS.d/next/Library/2022-03-20-17-15-56.bpo-47067.XXLnje.rst
@@ -1,0 +1,1 @@
+Implement :pep:`590` ``vectorcall`` for ``GenericAlias`` object