@corona10 2022-03-20
FYI, we decided to add caching mechanism instead of applying vector call. cc @sweeneyde 
see https://github.com/python/cpython/pull/19677#discussion_r413894982