@corona10 2022-03-20
@sweeneyde since @gvanrossum delegate decision to merge this PR or not, I will assign the PR to you :)
I am +1 about improving CPython performance if possible, so I have the same opinion with Guido too.