@gvanrossum 2022-03-20
> BTW, I replicated some speedup with pyperf commmands like
> 
> ```
> ./python -m pyperf timeit -s "D = dict[str, int]" "D(a=1, b=2)" --duplicate 10 -o "dict[str, int](a=1, b=2).json"
> ```
> 
> Some results:
> 
> ```python
> dict[str, int](a=1, b=2)      # 359 ns +- 6 ns -> 280 ns +- 3 ns: 1.29x faster
> list[int](())                 # 259 ns +- 3 ns -> 246 ns +- 5 ns: 1.05x faster
> MappingProxyType[str, int](d) # 273 ns +- 13 ns -> 261 ns +- 4 ns: 1.05x faster
> 
> class A:
>     def __init__(self, a, b):
>         pass
> G = GenericAlias(A, int)
> G(1, 2)                       # 198 ns +- 6 ns -> 190 ns +- 4 ns: 1.04x faster
> ```

So what does this speed up -- the `dict[str, int]` part, or the `...(a=1, b=2)` part? (Also, the latter is probably not the fastest way to create a dict. The 5% faster results are more in line with my expectations.