@sweeneyde 2022-03-20
> ...(a=1, b=2)

The `D(a=1, b=2)` part is what I benchmarked, for all of those types.