@gvanrossum 2022-03-21
If we’re now changing Makefile this PR seems to have strayed from the topic?