@sweeneyde 2022-03-21
> If we’re now changing Makefile this PR seems to have strayed from the topic?

I agree, we should probably keep the vectorcall changes for now, then explore global strings in another PR.