@sweeneyde 2022-03-21
Thanks for the changes, this is looking close.

I am still not convinced that adding `ga_make_tp_call` is worth it.

* It eliminates a few branches/C-calls, but doesn't eliminate an allocation or much data movement like `ga_vectorcall` does.
* Most other places in the code base, there is at most one vectorcall function per type, which is either present or NULL, so there's an argument to maintain consistency.
* This is not a particularly hot function: `list[int]`, `dict[str, int]`, `set[int]`, etc, all use the `ga_vectorcall` codepath, and `GenericAlias(MyPythonClass, some_type)()` will generally have lots of other overhead already.
* It requires using the private function `_PyObject_MakeTpCall` and grabbing the ThreadState, making the implementation less obvious and more tightly coupled.
* Three code paths is more complex than two.