@penguin-wwy 2022-03-21
In my mind, `ga_make_tp_call` will have an essential effect on tp_call, which still only needs to be packed once.

> It requires using the private function _PyObject_MakeTpCall and grabbing the ThreadState, making the implementation less obvious and more tightly coupled.

For the purpose of simplicity and decoupling, it is appropriate to delete it :)