@sweeneyde 2022-03-21
The string change seems to have really improved the microbenchmarks. Before/after this PR:
```
./python -m pyperf timeit -s "D = dict[str, int]" "D(a=1, b=2)" --duplicate 10 -o "dictbench.json" --rigorous
Mean +- std dev: [./dictbench.json] 366 ns +- 6 ns -> [../ga_vectorcall/dictbench.json] 202 ns +- 2 ns: 1.81x faster

./python -m pyperf timeit -s "L = list[int]" "L(())" --duplicate 10 -o "listbench.json" --rigorous
Mean +- std dev: [./listbench.json] 258 ns +- 5 ns -> [../ga_vectorcall/listbench.json] 168 ns +- 4 ns: 1.54x faster

./python -m pyperf timeit -s "from types import GenericAlias; PyClass = GenericAlias(type('A', (), {}), int)" "PyClass()" --duplicate 10 -o "pyclassbench.json" --rigorous
Mean +- std dev: [./pyclassbench.json] 153 ns +- 3 ns -> [../ga_vectorcall/pyclassbench.json] 75.9 ns +- 0.9 ns: 2.01x faster
```

I also checked for refleaks just for fun, and `./python -m test test_types test_typing test_genericalias -R3:3` passed.


Thanks for the contribution!