@gvanrossum 2022-03-21
Great results! In the end, how was the "make regen-all" issue resolved?