@sweeneyde 2022-03-21
> Great results! In the end, how was the "make regen-all" issue resolved?

I don't think it was solved, but ignored for the sake of this particular PR. I added a comment to https://bugs.python.org/issue46712 inquiring about it